+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added composer del/backspace guard to prevent accidental message delete
 - Fixed detached composer system theme detection - [#153](https://github.com/hkdb/aerion/issues/153)
 - Fixed launch flow - [#154](https://github.com/hkdb/aerion/issues/154)
+- Fixed dark theme rendering - [#155](https://github.com/hkdb/aerion/issues/155)
 
 
 **v0.2.2 - 05-09-2026**